chore: move LeftSidePanelContent to legacy
